+++ src/main/java/kr/pincoin/durian/auth/service/AdminService.java
@@ -3,7 +3,7 @@ package kr.pincoin.durian.auth.service;
 import kr.pincoin.durian.auth.domain.User;
 import kr.pincoin.durian.auth.domain.converter.Role;
 import kr.pincoin.durian.auth.domain.converter.UserStatus;
-import kr.pincoin.durian.auth.dto.UserCreateRequest;
+import kr.pincoin.durian.auth.controller.dto.UserCreateRequest;
 import kr.pincoin.durian.auth.repository.jpa.UserRepository;
 import kr.pincoin.durian.common.exception.ApiException;
 import lombok.RequiredArgsConstructor;
