+++ src/main/java/kr/pincoin/durian/auth/service/AuthService.java
@@ -3,10 +3,10 @@ package kr.pincoin.durian.auth.service;
 import kr.pincoin.durian.auth.domain.RefreshToken;
 import kr.pincoin.durian.auth.domain.User;
 import kr.pincoin.durian.auth.domain.converter.UserStatus;
-import kr.pincoin.durian.auth.dto.AccessTokenResponse;
-import kr.pincoin.durian.auth.dto.PasswordGrantRequest;
-import kr.pincoin.durian.auth.dto.RefreshTokenRequest;
-import kr.pincoin.durian.auth.dto.UserChangePasswordRequest;
+import kr.pincoin.durian.auth.controller.dto.AccessTokenResponse;
+import kr.pincoin.durian.auth.controller.dto.PasswordGrantRequest;
+import kr.pincoin.durian.auth.controller.dto.RefreshTokenRequest;
+import kr.pincoin.durian.auth.controller.dto.UserChangePasswordRequest;
 import kr.pincoin.durian.auth.jwt.TokenProvider;
 import kr.pincoin.durian.auth.repository.jpa.UserRepository;
 import kr.pincoin.durian.auth.repository.redis.RefreshTokenRepository;
