+++ src/main/java/kr/pincoin/durian/notifications/controller/dto/LineNotifyRequest.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.notifications.dto;
+package kr.pincoin.durian.notifications.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AccessLevel;
@@ -7,15 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MailgunSendResult {
-    @JsonProperty("id")
-    private String id;
-
+public class LineNotifyRequest {
     @JsonProperty("message")
     private String message;
-
-    public MailgunSendResult(String id, String message) {
-        this.id = id;
-        this.message = message;
-    }
 }
