+++ src/main/java/kr/pincoin/durian/notifications/service/AligoService.java
@@ -2,9 +2,9 @@ package kr.pincoin.durian.notifications.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import kr.pincoin.durian.notifications.dto.AligoSendRequest;
-import kr.pincoin.durian.notifications.dto.AligoSendResponse;
-import kr.pincoin.durian.notifications.dto.AligoSendResult;
+import kr.pincoin.durian.notifications.controller.dto.AligoSendRequest;
+import kr.pincoin.durian.notifications.controller.dto.AligoSendResponse;
+import kr.pincoin.durian.notifications.service.dto.AligoSendResult;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
