+++ src/main/java/kr/pincoin/durian/notifications/service/dto/MailgunSendResult.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.notifications.dto;
+package kr.pincoin.durian.notifications.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AccessLevel;
@@ -7,7 +7,15 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LineNotifyRequest {
+public class MailgunSendResult {
+    @JsonProperty("id")
+    private String id;
+
     @JsonProperty("message")
     private String message;
+
+    public MailgunSendResult(String id, String message) {
+        this.id = id;
+        this.message = message;
+    }
 }
