+++ src/main/java/kr/pincoin/durian/shop/controller/CategoryController.java
@@ -3,8 +3,8 @@ package kr.pincoin.durian.shop.controller;
 import jakarta.validation.Valid;
 import kr.pincoin.durian.common.exception.ApiException;
 import kr.pincoin.durian.shop.domain.conveter.CategoryStatus;
-import kr.pincoin.durian.shop.dto.CategoryCreateRequest;
-import kr.pincoin.durian.shop.dto.CategoryResponse;
+import kr.pincoin.durian.shop.controller.dto.CategoryCreateRequest;
+import kr.pincoin.durian.shop.controller.dto.CategoryResponse;
 import kr.pincoin.durian.shop.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
