+++ src/main/java/kr/pincoin/durian/shop/controller/dto/ProductCreateRequest.java
@@ -1,0 +1,4 @@
+package kr.pincoin.durian.shop.controller.dto;
+
+public class ProductCreateRequest {
+}
