--- src/main/java/kr/pincoin/durian/shop/dto/ProductCreateRequest.java
@@ -1,4 +1,0 @@
-package kr.pincoin.durian.shop.dto;
-
-public class ProductCreateRequest {
-}
