--- src/main/java/kr/pincoin/durian/shop/dto/ProductResponse.java
@@ -1,4 +1,0 @@
-package kr.pincoin.durian.shop.dto;
-
-public class ProductResponse {
-}
